+++ server-side/tests_functions.ts
@@ -1,3 +1,4 @@
+import { SchemasRequiringSyncTests } from '../server-side/tests/api_tests/SchemasRequiringSyncTests.test';
 import { DataIndex } from './tests/api_tests/DataIndex.test';
 import { NebulaTest } from './tests/api_tests/NebulaTest.test';
 import { NebulaTestPart2 } from './tests/api_tests/NebulaTestPart2.test';
@@ -205,3 +206,15 @@ export async function nebula_internal_test(client: Client, addonClient: Client, 
     return (await testerFunctions.run());
 };
 context["nebula_internal_test"] = nebula_internal_test;
+
+export async function schemas_requiring_sync_tests(client: Client, addonClient: Client, request: Request, testerFunctions: TesterFunctions) {
+    const service = new GeneralService(client);
+    const serviceAddon = new GeneralService(addonClient);
+    testName = 'SchemasRequiringSyncTests'; //printing your test name - done for logging
+    service.PrintMemoryUseToLog('Start', testName);
+    testerFunctions = service.initiateTesterFunctions(client, testName);
+    await SchemasRequiringSyncTests(service, serviceAddon, request, testerFunctions);//this is the call to YOUR test function
+    await test_data(client, testerFunctions);//this is done to print versions at the end of test - can be deleted
+    return (await testerFunctions.run());
+};
+context["schemas_requiring_sync_tests"] = schemas_requiring_sync_tests;
